requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
selenium~=4.11.2
celery~=5.3.1
redis~=5.0.0
django-celery-beat~=2.5.0
persiantools~=3.0.1
Unidecode~=1.3.6